Fix ColorSlider hue gradient
The background gradient used by the hue slider of ColorPicker does not
accurately represent a gradient through hues with constant saturation/
lightness because it is implemented with 10 stops instead of the 6
primary/secondary stops needed to correctly represent the hue gradient
with an RGB interpolation.

In particular, the current gradient has incorrect lightness dips around
fuchsia and yellow and (less noticeable) lightness bumps around lime and
blue.